Seperated algorithms into modules with corresponding test files
